Created WorldScreen class and moved relevant methods out of the Roam class and into it.
